Patched an Issue
This Version patches the unlimited "Bot is thinking" when running the checkorder command. The issue that caused this was that the value of some variables just was an empty list `[]` what discord could not send in an embed. i edited it so if the value of these variables is `[]` it change the value to `Not found` and this works fine at the moment.